requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.101.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
jwt==1.3.1
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydentic==0.0.1.dev3
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-stdnum==1.18
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.7
